reformated
added new files to separate and clean-up code for practice on larger projects
