+++ src/App.jsx
@@ -1,23 +1,22 @@
 import { useState } from "react"
 import "./styles.css"
+import { NewTodoForm } from "./NewTodoForm"
+import { TodoList } from "./TodoList"
 
 export default function App() {
   const [newItem, setNewItem] = useState("")
   const [todos, setTodos] = useState([])
 
-  function handleSubmit(e) {
-    e.preventDefault()
-
+  function addTodo(title){
     setTodos(currentTodos => {
-      return[
-        ...currentTodos, 
-        {id: crypto.randomUUID(), 
-        title: newItem, 
-        completed: false}
-      ]
+    return[
+    ...currentTodos, 
+    {id: crypto.randomUUID(), 
+    title, 
+    completed: false}
+    ]
 
     })
-    setNewItem("")
   }
 
   function toggleTodo(id, completed){
@@ -39,34 +38,11 @@ export default function App() {
 
   return(
   <>
-  <form onSubmit={handleSubmit} className="new-item-form">
-    <div className="form-row">
-      <label htmlFor="item">New Item</label>
-      <input 
-      value={newItem}
-      onChange={e => setNewItem(e.target.value)}
-      type="text" 
-      id="item" />
-    </div>
-    <button className="btn">Add</button>
-  </form>
+  <NewTodoForm onSubmit={addTodo}/>
   <h1 className="header">ToDo List</h1>
-  <ul className="list">
-    {todos.length === 0 && "No Todos"}
-    {todos.map(todo => {
-      return (<li key={todo.id}>
-      <label>
-        <input type="checkbox" 
-        checked = {todo.completed}
-        onChange={e => toggleTodo(todo.id, e.target.checked)} />
-        {todo.title}
-      </label>
-      <button onClick={() => deleteTodo(todo.id)}
-       className="btn btn-danger"
-       >Delete</button>
-    </li>
-    )})}
-  </ul>
+  <TodoList todos={todos}
+   toggleTodo={toggleTodo}
+   deleteTodo={deleteTodo} />
   </>
   )
 }
