+++ src/TodoItem.jsx
@@ -1,0 +1,17 @@
+export function TodoItem({completed, id, title, toggleTodo, deleteTodo}){
+    return(
+        <li>
+        <label>
+          <input type="checkbox" 
+          checked = {completed}
+          onChange={e => toggleTodo(id, e.target.checked)}
+          />
+          {title}
+        </label>
+        <button 
+         onClick={() => deleteTodo(id)}
+         className="btn btn-danger"
+         >Delete</button>
+      </li>
+    )
+}
